% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_relationship.R
\name{access_reactome}
\alias{access_reactome}
\title{Return data from reactome}
\usage{
access_reactome()
}
\value{
A list of paths and human genes in the list
}
\description{
Looks in reactome for the pathways associated with each gene
}
\examples{
paths2genes <- access_reactome()
}
