% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{allComb}
\alias{allComb}
\title{Logical vectors of meta data}
\usage{
allComb(data, columns)
}
\arguments{
\item{data}{A \code{data.frame} with factors}

\item{columns}{names of the columns to be used}
}
\value{
a matrix with the logical values of each combination of the levels of
the columns given for the data in each column.
}
\description{
Given the names of the columns of the data calculates the logical vectors of
each subset of the data
}
\note{
If some rows are all FALSE it means some values are NA.
}
