% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{angle}
\alias{angle}
\title{Calculates the angle between to slopes}
\usage{
angle(x, y = 1)
}
\arguments{
\item{x, y}{Slope of the lines}
}
\value{
The smaller angle between the slopes in degrees
}
\description{
Calculates the angle between to slopes
}
\note{
The default compares the first slope with the slope 1
}
