% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RGCCA.R
\name{boot_evaluate}
\alias{boot_evaluate}
\title{Evaluates the boostrapping of RGCCA}
\usage{
boot_evaluate(STAB)
}
\arguments{
\item{STAB}{List of weights of \code{rgcca} or \code{sgcca}}
}
\value{
Lateral effect: Prints plots
}
\description{
Evaluates the boostrapping of RGCCA
}
