% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{boots_corr}
\alias{boots_corr}
\title{Bootstrapping to asses how probable is to have such enrichment in those p-values matrix}
\usage{
boots_corr(pvalues, component, iter = 10000)
}
\arguments{
\item{pvalues}{Vector of pvalues}

\item{component}{The named component different from 0 from a SGCCA object}

\item{iter}{The number of permutations}
}
\value{
The numeric pvalue of the bootstrap.
}
\description{
Bootstrapping to asses how probable is to have such enrichment in those p-values matrix
}
