% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RGCCA.R
\name{cca_rgcca}
\alias{cca_rgcca}
\title{Calculate correlation and covariance between CCA dimensions}
\usage{
cca_rgcca(rgcca)
}
\arguments{
\item{rgcca}{The output of SGCCA or RGCCA}
}
\value{
A list of matrix with the correlation and covariation between CCA
dimensions
}
\description{
Calculate correlation and covariance between CCA dimensions
}
