% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design.R
\name{check_design}
\alias{check_design}
\title{Validate designs}
\usage{
check_design(designs)
}
\arguments{
\item{designs}{A list of design matrices as obtained from
\code{\link{weight_design}}}
}
\value{
A logical vector that validates if the designs are correct or not.
}
\description{
Validate designs
}
\examples{
designs <- weight_design(4, 4)
keep <- check_design(designs)
summary(keep)
}
\seealso{
\code{\link{symm}}, \code{\link{subSymm}}
}
