% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{circle}
\alias{circle}
\title{A circle of radium 2}
\format{
An object of class \code{data.frame} with 100 rows and 2 columns.
}
\usage{
circle
}
\description{
Created with \code{circleFun}
}
\keyword{datasets}
