% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{circleFun}
\alias{circleFun}
\title{Create a circle}
\usage{
circleFun(center = c(-1, 1), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{The position where the center of the circle is}

\item{diameter}{Arbitrary units of diameter of the circle}

\item{npoints}{Number of points of the circle (aka: definition of the circle)}
}
\value{
A data.frame with the position of the points
}
\description{
Create a circle
}
