% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{clean_unvariable}
\alias{clean_unvariable}
\title{Remove unvariable features}
\usage{
clean_unvariable(A)
}
\arguments{
\item{A}{A list of matrices with variables as columns.}
}
\description{
Remove variables with sd equal to 0
}
