% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prevalence.R
\name{comb_prevalence}
\alias{comb_prevalence}
\title{Look for prevalence in combinations of two}
\usage{
comb_prevalence(table, meta, columns)
}
\arguments{
\item{table}{is the data}

\item{meta}{is the metadata}

\item{columns}{is the columns to make comparisons from}
}
\value{
A table with all the pairwise comparisons
}
\description{
Look for prevalence in combinations of two
}
\seealso{
\code{\link[=prevalence]{prevalence()}}, \code{\link[=prevalence_tab]{prevalence_tab()}}, \code{\link[=full_prevalence]{full_prevalence()}}
}
