% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prevalence.R
\name{comp}
\alias{comp}
\title{Analyze the data for the relationship in time}
\usage{
comp(input, meta)
}
\arguments{
\item{input}{the table of microorganisms}

\item{meta}{the metadata it has encoded things only for intestinal}
}
\value{
A list of the p-values the first two are for all comparisons
the following are for ileum and colon.
}
\description{
It looks in general and in colon and ILEUM specifics
}
