% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compare.correlations}
\alias{compare.correlations}
\title{Calculates the z-score of two correlations}
\usage{
compare.correlations(r1, r2, n1, n2)
}
\arguments{
\item{r1, r2}{The correlation coefficients}

\item{n1, n2}{The number of samples used to calculate the correlations}
}
\value{
The z score of the comparison of the coefficients.
}
\description{
Calculates the z-score of two correlations
}
