% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compares two objects of class sgcca}
\usage{
compare(object1, object2)
}
\arguments{
\item{object1, object2}{sgcca objects}
}
\value{
Genes In common
}
\description{
Returns the number of genes in common, it assumes the first componenet is
always the genes.
}
