% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{contingency_taxa}
\alias{contingency_taxa}
\title{Compares the taxonomy of the otus}
\usage{
contingency_taxa(taxa_1, taxa_2)
}
\arguments{
\item{taxa_1, }{taxa_2 taxonomic table as described in taxonomy}
}
\value{
A matrix of nrow(taxa_1) * nrow(taxa_2)
}
\description{
Given two taxonomy tables find which one is in which one
}
