% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert.z.score}
\alias{convert.z.score}
\title{Z-score to p-value}
\usage{
convert.z.score(z, one.sided = NULL)
}
\arguments{
\item{z}{the normalized value}

\item{one.sided}{Either NULL, - or +}
}
\value{
The p-value
}
\description{
Calculates the p-value of a z-score
}
