% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sign.R
\name{cor_sign}
\alias{cor_sign}
\title{Significative correlation}
\usage{
cor_sign(n)
}
\arguments{
\item{n}{The number of data points where the correlation has been done}
}
\value{
the correlation value at which it becomes below 0.05
}
\description{
Calculates the correlation at which they become significant
}
