% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RGCCA.R
\name{cors_rgcca}
\alias{cors_rgcca}
\title{Check the efficacy of RGCCA}
\usage{
cors_rgcca(a, rgcca)
}
\arguments{
\item{a}{The value}

\item{rgcca}{The output of \code{\link[RGCCA:sgcca]{RGCCA::sgcca()}} or
\code{\link[RGCCA:rgcca]{RGCCA::rgcca()}}}
}
\description{
This function test some help from
\url{https://onlinecourses.science.psu.edu/stat505/node/68}
Performs the correlation between the original variables and the
resulting components (of each block). To check if the
}
