% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{ensembl2symbol}
\alias{ensembl2symbol}
\title{Translate ensmbl to symbols}
\usage{
ensembl2symbol(x)
}
\arguments{
\item{x}{A data.frame with a column named Gene}
}
\value{
The same data.frame with gene names without duplications and empty
symbols
}
\description{
Translate ensmbl to symbols
}
