% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epithelium.R
\name{epitheliumE}
\alias{epitheliumE}
\title{Function to read the epithelium signature}
\usage{
epitheliumE(x)
}
\arguments{
\item{x}{A character vector with SYMBOL names}
}
\value{
A character vector with the Entrez ids
}
\description{
Return the Epithelium in Entrez ids
}
