% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fastercheck}
\alias{fastercheck}
\title{Check if a vector is in the matrix}
\usage{
fastercheck(x, matrix)
}
\arguments{
\item{x}{The matrix to check if it is in the matrix}

\item{matrix}{The matrix where x is looked up.}
}
\value{
A logical vector saying if the vector is in the matrix
}
\description{
Check if a vector is in the matrix
}
