% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_RNAseq}
\alias{filter_RNAseq}
\title{Filter expressions}
\usage{
filter_RNAseq(expr, frac = 0.1)
}
\arguments{
\item{expr}{Input RNAseq data (not microbiota)}

\item{frac}{The fraction to remove}
}
\value{
A matrix without low expressed genes and with low variance
}
\description{
Filter expressions
}
\section{Remove the expression that is not in more than 25\% of samples and the lower}{
10\% of genes
}

