% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{filter_values}
\alias{filter_values}
\title{Filter correlations}
\usage{
filter_values(comp1, cors, pval, threshold)
}
\arguments{
\item{comp1}{A component with names}

\item{cors}{A matrix with the correlations}

\item{pval}{A matrix with the pvalues}

\item{threshold}{The numeric threshold of the filter}
}
\value{
A list with matrices of correlations and pvalues that pass the filter
}
\description{
Filter the correlations based on the pvalue
}
\seealso{
\code{\link[=sign_cor]{sign_cor()}}, \code{\link[=relevant]{relevant()}}
}
