% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prevalence.R
\name{full_prevalence}
\alias{full_prevalence}
\title{Look for prevalence of a factor}
\usage{
full_prevalence(table, meta, column)
}
\arguments{
\item{table}{is the data}

\item{meta}{A data.frame from which column is is the metadata}

\item{column}{is the column to see if is associated with the presence or absence.}
}
\value{
A data.frame with all the pairwise comparisons
}
\description{
Look for prevalence of a factor
}
\seealso{
\code{\link[=prevalence]{prevalence()}}, \code{\link[=prevalence_tab]{prevalence_tab()}}, \code{\link[=full_prevalence]{full_prevalence()}}
}
