% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing-models.R
\name{getAVEs}
\alias{getAVEs}
\title{Retrieve inner AVE}
\usage{
getAVEs(x, pos = 1L)
}
\arguments{
\item{x}{The sgcca or rgcca object}

\item{pos}{The position of the AVE_inner to extract}
}
\value{
A numeric value
}
\description{
Retrieve inner AVE
}
