% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_correlations.R
\name{intercorrelation}
\alias{intercorrelation}
\title{Intracorrelation}
\usage{
intercorrelation(A, ids)
}
\arguments{
\item{A}{Matrix}

\item{ids}{Vector of ids of the sample}
}
\value{
the pearson coefficient
}
\description{
Calculates the intercorrelation between ids
(the correlation between different ids)
}
