% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_correlations.R
\name{intracorrelation}
\alias{intracorrelation}
\title{Intracorrelation}
\usage{
intracorrelation(A, ids)
}
\arguments{
\item{A}{Matrix}

\item{ids}{Vector of ids of the sample}
}
\value{
the pearson coefficient
}
\description{
Calculates the intracorrelation between ids
}
