% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{kFolding}
\alias{kFolding}
\title{Indexes for K-folds}
\usage{
kFolding(k, size, n, ...)
}
\arguments{
\item{k}{Number of k-folds (two divides the dataset in 2)}

\item{size}{The number of samples required}

\item{n}{Number of k-folds to do}

\item{...}{Other arguments passed to sample the}
}
\value{
An list of length \code{n} with two elements the training and the
testing elements with the index of the samples
}
\description{
Calculates the indices of the k-fold bootstrapping for other functions
}
\seealso{
\code{\link{looIndex}}
}
