% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design.R
\name{loo_functions}
\alias{loo_functions}
\title{Function to export a function}
\usage{
loo_functions(A, shrinkage)
}
\arguments{
\item{A}{The original list of data}

\item{shrinkage}{The shrinkage parameter}
}
\value{
A function with two arguments:
\subsection{x}{the index}
\subsection{model}{the model used}
}
\description{
Function to export a function
}
