% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing-models.R
\name{m_sem}
\alias{m_sem}
\title{Summarize a model}
\usage{
m_sem(model, loo)
}
\arguments{
\item{model}{Original sgcca object}

\item{loo}{A list with using the same model but using a leave-one-out
strategy.}
}
\value{
The inner AVE, the Mean and the SEM of the \code{loo}
}
\description{
Summarize a model
}
\seealso{
\code{\link[=sem]{sem()}}
}
