% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{makeRects}
\alias{makeRects}
\title{Make a rectangle at those cells}
\usage{
makeRects(tfMat, border)
}
\arguments{
\item{tfMat}{matrix to take the dimensions from and with TRUE at the cells
to be higlighted}

\item{border}{Type of border}
}
\value{
A side effect of the adding a rectangle to the heatmap
}
\description{
Make a rectangle at those cells
}
\references{
https://stackoverflow.com/a/45891116/2886003
https://stackoverflow.com/a/7980349/2886003
}
