% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RGCCA.R
\name{McKeonHomeogenity}
\alias{McKeonHomeogenity}
\title{Calculates McKeon Homeogenity}
\usage{
McKeonHomeogenity(B, C)
}
\arguments{
\item{B}{The list of blocks}

\item{C}{The design matrix}
}
\value{
A matrix of the same dimensions as the list and design matrix with
the "correlations" between blocks.
}
\description{
Looks how much each blocks is related to the others blocks. This shouldn't
be used with the metablock data (it is not a real block)
}
\references{
There are no references other than the vignette of RGCCA
}
