% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{meta_i_norm}
\alias{meta_i_norm}
\title{Normalize 16S biopsies metadata}
\usage{
meta_i_norm(meta)
}
\arguments{
\item{meta}{The metadata}
}
\value{
The dataframe with the metadata corrected
}
\description{
Normalize 16S biopsies metadata
}
