% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{meta_r_norm}
\alias{meta_r_norm}
\title{Normalize the metadata of the RNA}
\usage{
meta_r_norm(meta)
}
\arguments{
\item{meta}{Meta}
}
\value{
A data.frame with some modifications
}
\description{
Performs some modifications to the data.frame
}
