% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{meta_s_norm}
\alias{meta_s_norm}
\title{Normalize 16S stools metadata}
\usage{
meta_s_norm(meta)
}
\arguments{
\item{meta}{The metadata}
}
\value{
The dataframe with the metadata corrected
}
\description{
Normalize 16S stools metadata
}
