% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design.R
\name{model_columns}
\alias{model_columns}
\title{Adapt data for a CCA}
\usage{
model_columns(data, columns)
}
\arguments{
\item{data}{A data.frame with the information about the samples}

\item{columns}{The name of the columns to be used to build the matrix}
}
\value{
A matrix
}
\description{
Convert factors to numeric (in order of appearance), the numeric variables
are left as is.
}
\note{
The \code{NA} are converted to 0.
}
\seealso{
\code{\link{model_RGCCA}}
}
