% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design.R
\name{model_RGCCA}
\alias{model_RGCCA}
\title{Prepare data for CCA.}
\usage{
model_RGCCA(data, columns, intercept = FALSE)
}
\arguments{
\item{data}{A data.frame with the information about the samples}

\item{columns}{The name of the columns to be used to build the matrix}

\item{intercept}{A logical value if you want one column with all 1 or not.}
}
\value{
A matrix with each factor is decomposed in as much columns as
factors has minus 1 and with the numeric values as they were.
}
\description{
Prepares the factors into their vectors. Each level of a factor is converted
to a column, numeric columns are left as is.
}
\note{
The \code{NA} are converted to 0
}
\seealso{
\code{\link{model_columns}}
}
