% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{norm_expr_colnames}
\alias{norm_expr_colnames}
\title{Change names of expression}
\usage{
norm_expr_colnames(expr)
}
\arguments{
\item{expr}{The expression matrix}
}
\value{
The same matrix with only modifications on the colnames
}
\description{
Swap some samples, correct the names of a pair of samples
}
