% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{norm_otus}
\alias{norm_otus}
\title{Normalize OTUS}
\usage{
norm_otus(otus_table_i, otus_tax_i = NULL)
}
\arguments{
\item{otus_table_i}{OTUs table}

\item{otus_tax_i}{Taxonomic table}
}
\value{
A normalized and filtered matrix
}
\description{
Uses metagenomeSeq to normalize and filter if sd is 0
}
