% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{norm_RNAseq}
\alias{norm_RNAseq}
\title{Normalize RNAseq}
\usage{
norm_RNAseq(expr)
}
\arguments{
\item{expr}{A matrix of genes and samples}
}
\value{
A normalized matrix with TMM
}
\description{
Uses edgeR to normalize the expression data
}
