% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{pathsPerMicro}
\alias{pathsPerMicro}
\title{Enrichment by microorganisms}
\usage{
pathsPerMicro(x, all_genes)
}
\arguments{
\item{x}{The table of microoganisms, genes, correlations and adjusted p-values.}

\item{all_genes}{All the genes that are found relevant (weight != 0)}
}
\value{
A list of tables with the pathways enriched for every microorganism
}
\description{
Function to split by microorganism and compare if they are enriched in some
pathways
}
