% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovas.R
\name{permanova_otus}
\alias{permanova_otus}
\title{Permanova}
\usage{
permanova_otus(otus, meta)
}
\arguments{
\item{otus}{Matrix of the variables to analyze}

\item{meta}{Data frame with the variables to compare}
}
\value{
An anova.cca object
}
\description{
Permanova
}
