% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot a correlation with ggplot}
\usage{
plot_cor(file, cors, pval, threshold, label)
}
\arguments{
\item{file}{The name of the file with the model}

\item{cors}{A matrix with the correlations}

\item{pval}{A matrix with the pvalues}

\item{threshold}{The numeric threshold of the filter}

\item{label}{Label for the output html file}
}
\value{
Create a html file at Figures/heatmap...html
}
\description{
Use filter_values to decide how to and create an html file
}
\note{
Expects genes in rows and species at the columns in \code{cors} and
\code{pvalue}
}
