% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_interesting}
\alias{plot_interesting}
\title{Plot  PCA of interesting variables}
\usage{
plot_interesting(subVariables, meta_r, expr, otus_table_i)
}
\arguments{
\item{subVariables}{A data.frame}

\item{meta_r}{The data.frame with the information about the samples}

\item{expr}{A matrix with the RNAseq expression}

\item{otus_table_i}{A matrix with the 16S}
}
\value{
A
}
\description{
Plot  PCA of interesting variables
}
