% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_samples}
\alias{plot_samples}
\title{Plot samples}
\usage{
plot_samples(samples, colors, individual = FALSE)
}
\arguments{
\item{samples}{A data.frame with RNAseq, Micro, labels columns}

\item{colors}{Manual colors for some samples}

\item{individual}{Logical. Do each plot by ID and Time?}
}
\value{
Several plots
}
\description{
Make several plots of the samples, using several colors and things
}
