% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{plot_single_cor}
\alias{plot_single_cor}
\title{Plot correlations}
\usage{
plot_single_cor(x, gene, y, microorganism, colr, case, cor_val)
}
\arguments{
\item{x}{Gene expression matrix}

\item{gene}{Name of the gene to plot}

\item{y}{Microorganism matrix}

\item{microorganism}{Name of the microorganism}

\item{colr}{Factor to use for colors}

\item{case}{Factor to use for points shape}

\item{cor_val}{Title of the correlation (usually the correlation value)}
}
\description{
Plot correlations
}
