% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Test prevalence}
\usage{
prevalence(presence, absence)
}
\arguments{
\item{presence}{Matrix with presence of certain microorganism}

\item{absence}{Matrix with the absence of certain microorganism}
}
\description{
Iterate in the tables to reach
}
\examples{
presence <- structure(c(13, 9, 8, 0, 1, 10, 8, 13, 0, 2),
                      .Dim = c(5L, 2L),
                      .Dimnames = list(NULL, c("0", "14")))

absence <- structure(c(29, 33, 34, 42, 41, 31, 33, 28, 41, 39),
                     .Dim = c(5L, 2L), .Dimnames = list(NULL, c("0", "14")))
prevalence(presence, absence)
}
\seealso{
\code{\link[=comb_prevalence]{comb_prevalence()}}, \code{\link[=prevalence_tab]{prevalence_tab()}}
}
