% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prevalence.R
\name{prevalence_tab}
\alias{prevalence_tab}
\title{Calculates the presence or absence of a microorganism}
\usage{
prevalence_tab(table, meta, columns)
}
\arguments{
\item{table}{Input data with the samples in columns and microorganism in rows}

\item{meta}{Metadata associated with those samples (it assumes they are in
the same order as in table)}

\item{columns}{Column to make contrasts from}
}
\value{
a list with the matrices of presence and absence
}
\description{
It takes into account only samples with more than 0.5% of relative presence
of the microorganism
}
