% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RGCCA.R
\name{probability_samples}
\alias{probability_samples}
\title{Calculates the probability of obtaining these samples.}
\usage{
probability_samples(meta)
}
\arguments{
\item{meta}{The data.frame where each column is a variable and the row is a
sample}
}
\value{
A numeric vector with the probability for each sample.
}
\description{
Given a data.frame with categories it looks how probable is to have such a
sample
}
\details{
If a row for a variable is \code{NA} it uses the median for that
variable
}
\examples{
probability_samples(iris[, c("Petal.Width", "Species")])
}
