% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sign.R
\name{pvalue}
\alias{pvalue}
\title{Calculates the p.value}
\usage{
pvalue(r, n)
}
\arguments{
\item{r}{Correlation coefficient}

\item{n}{Number of samples}
}
\value{
the p-value
}
\description{
Calculates the p.value
}
\seealso{
Other pvalue correlations: \code{\link{cor_sign}}
}
\concept{pvalue correlations}
