% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prevalence.R
\name{ratio}
\alias{ratio}
\title{Ratio of prevalence}
\usage{
ratio(columns, data, indices, meta)
}
\arguments{
\item{columns}{Columns to select}

\item{data}{Data.frame from which the columns are selected}

\item{indices}{Rows where we are interested in}

\item{meta}{# Information about the samples}
}
\description{
Calculates the ratio of p-values of the prevalence between samples
}
