% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{readSGCCA}
\alias{readSGCCA}
\title{Read component}
\usage{
readSGCCA(file, set = 1, component = 1)
}
\arguments{
\item{file}{The .RDS file were the sgcca object is stored}

\item{set}{The set for which we want the component}

\item{component}{The number of the component we want.}
}
\value{
The vector of the component without those variables with weight 0
}
\description{
Read component
}
