% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{relevant}
\alias{relevant}
\title{List the correlations}
\usage{
relevant(comp1, cors, pval, threshold = 0.05)
}
\arguments{
\item{comp1}{A component with names}

\item{cors}{A matrix with the correlations}

\item{pval}{A matrix with the pvalues}

\item{threshold}{The numeric threshold of the filter}
}
\value{
A data.frame with microorganisms, genes, their correlations and the
pvalue
}
\description{
Prepare the correlations to be shared
}
\seealso{
\code{\link[=filter_values]{filter_values()}}, \code{\link[=sign_cor]{sign_cor()}}
}
