% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{select_genes_int}
\alias{select_genes_int}
\title{Filter genes}
\usage{
select_genes_int(file, expr)
}
\arguments{
\item{file}{Name of the file with folder}

\item{expr}{Matrix of expression}
}
\value{
A Matrix with the filtered genes
}
\description{
Filter genes
}
