% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{select_var}
\alias{select_var}
\title{Select important variables}
\usage{
select_var(sgcca)
}
\arguments{
\item{sgcca}{result of sgcca or rgcca}
}
\value{
a data.frame with variables and their location
}
\description{
Select important variables
}
