% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{selectVar}
\alias{selectVar}
\title{Select variable from bootstrapping}
\usage{
selectVar(x)
}
\arguments{
\item{x}{List of the summary statistics of the bootstrapping}
}
\value{
The names of the selected variables
}
\description{
Select variable from bootstrapping
}
