% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{sign_cor}
\alias{sign_cor}
\title{List the correlations}
\usage{
sign_cor(cors, pval, threshold = 0.05)
}
\arguments{
\item{cors}{A matrix with the correlations}

\item{pval}{A matrix with the pvalues}

\item{threshold}{The numeric threshold of the filter}
}
\value{
A data.frame with microorganisms, genes, their correlations and the
pvalue
}
\description{
Prepare the correlations to be shared (without filtering them)
}
\seealso{
\code{\link{filter_values}}, \code{\link{relevant}}
}
