% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{size}
\alias{size}
\title{Number of samples}
\usage{
size(x)
}
\arguments{
\item{x}{A list or a matrix where rows are the samples}
}
\value{
The number of samples
}
\description{
Number of samples
}
