% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{store_micro}
\alias{store_micro}
\title{Store result by microorganism}
\usage{
store_micro(x, o2, label)
}
\arguments{
\item{x}{Name of microorganism}

\item{o2}{Filter samples}

\item{label}{Label for the files}
}
\description{
Store result by microorganism
}
