% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{subsetData}
\alias{subsetData}
\title{Subset a list}
\usage{
subsetData(A, index)
}
\arguments{
\item{A}{A list of an array with samples in rows.}

\item{index}{The number of samples to keep}
}
\value{
A list with less samples
}
\description{
Given a list and an index subset each element of the list.
}
