% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design.R
\name{symm}
\alias{symm}
\title{Create symmetric matrix}
\usage{
symm(m, data)
}
\arguments{
\item{m}{Square matrix.}

\item{data}{Numeric values of the upper triangular side of the matrix}
}
\value{
A square symmetric matrix.
}
\description{
Create symmetric matrix
}
