% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{taxonomy}
\alias{taxonomy}
\title{Clean and prepare the data from IMNGS}
\usage{
taxonomy(taxonomy, otus)
}
\arguments{
\item{taxonomy}{Last column of files a string ; separated with domain,
phylum, vlass, order, family, genus and species.}

\item{otus}{The name of the rows}
}
\value{
A matrix with the taxonomic information ready for the package phylo
}
\description{
Divides the taxonomy into a new matrix for each otu
}
