% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing-models.R
\name{tidyer}
\alias{tidyer}
\title{Clean the output of a sgcca object}
\usage{
tidyer(data, model, type)
}
\arguments{
\item{data}{The matrices of \code{sggcca$Y[[1]]} or \code{sgcca$a[[1]]}}

\item{model}{A label of the model used}

\item{type}{A label to know what it}
}
\value{
A \code{data.frame}
}
\description{
Clean the output of a sgcca object
}
