% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{tol21rainbow}
\alias{tol21rainbow}
\title{Rainbow colors}
\format{An object of class \code{character} of length 21.}
\usage{
tol21rainbow
}
\description{
Created or obtained from an online resource
}
\keyword{datasets}
