% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimVer.R
\name{trimVer}
\alias{trimVer}
\title{Trim version number of genes}
\usage{
trimVer(genes)
}
\arguments{
\item{genes}{A character vector.}
}
\value{
A vector of genes names
}
\description{
Remove the trailing version of gene names in the ENSEMBEL versioning
scheme.
}
\examples{
trimVer("ENSG00000049192.14")
trimVer(c("ENSG00000049192.14", "ENSG00000065320.8"))
}
