% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{two.sided}
\alias{two.sided}
\title{Two sided test}
\usage{
two.sided(y, z)
}
\arguments{
\item{y}{Value of the test}

\item{z}{Vector of of the permutations}
}
\value{
The p-value
}
\description{
Test in a vector from a permutation if there is a relationship or not.
Assumes that the distribution is symmetric around 0.
}
