% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sgcca.R
\name{variables}
\alias{variables}
\title{Extract in a tidy way the information about the variables}
\usage{
variables(sgcca)
}
\arguments{
\item{sgcca}{A \code{sgcca} object}
}
\value{
A data.frame of variables, component and origin.
}
\description{
Extract in a tidy way the information about the variables
}
