% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RGCCA.R
\name{variables_weight}
\alias{variables_weight}
\title{Plot density of the weight of components}
\usage{
variables_weight(comp)
}
\arguments{
\item{comp}{Component from \code{sapply(rgcca$a, function(x)x[, 1])}}
}
\value{
Lateral effect: A plot, invisible the ggplot object of the plot
}
\description{
Plot density of the weight of components
}
