% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_design.R
\name{weight_design}
\alias{weight_design}
\title{Check a design for different weights}
\usage{
weight_design(weights = 4, size)
}
\arguments{
\item{weights}{A numeric value with the number of weights to be used}

\item{size}{A numveric value with the numer of datasets on the design.}
}
\value{
A list of matrices with the designs with different weights
}
\description{
Check a design for different weights
}
\examples{
weight_design(4, 4)
}
