% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{weights}
\alias{weights}
\title{Select the weights and adds information for human RNAseq}
\usage{
weights(x)
}
\arguments{
\item{x}{a RGCCA outputs}
}
\value{
a data.frame
}
\description{
Select the weights and adds information for human RNAseq
}
