% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_relations.R
\name{weights_bayes}
\alias{weights_bayes}
\title{Independence between genes and OTUs}
\usage{
weights_bayes(x, absolute = FALSE)
}
\arguments{
\item{x}{A list containing sgcc objects}

\item{absolute}{A logical value to check if the weight should be taken in
absolute value or not}
}
\value{
A matrix of p-values of the fisher test to evaluate the independence
of the pair.
}
\description{
Independence between genes and OTUs
}
\note{
We want related pairs, so the higher the p-value, the better.
}
