% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_relations.R
\name{weights_correlation}
\alias{weights_correlation}
\title{Correlation weights}
\usage{
weights_correlation(x)
}
\arguments{
\item{x}{A list with different sgcca objects (usually from the leave-one-out
procedure)}
}
\value{
A matrix with correlations of genes and OTUs
}
\description{
Select those variables with some weights, filter out those that appear in
less than half the list, and make the correlation (using spearman)
}
