% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{weights_otus}
\alias{weights_otus}
\title{Select the weights and adds information for taxa RNAseq}
\usage{
weights_otus(x, taxa)
}
\arguments{
\item{x}{a RGCCA outputs}

\item{taxa}{The taxonomic levels}
}
\value{
a data.frame
}
\description{
Select the weights and adds information for taxa RNAseq
}
