% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{write_cor}
\alias{write_cor}
\title{Translate symbols and store}
\usage{
write_cor(x, file)
}
\arguments{
\item{x}{A data.frame with a column named Gene}

\item{file}{The name of the file were to store them}
}
\description{
Translate symbols and store
}
\seealso{
\code{\link{ensembl2symbol}}
}
