% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_relationship.R
\name{biological_relationships}
\alias{biological_relationships}
\title{Performs the calculation of biological information}
\usage{
biological_relationships(
  sgcca.centroid,
  STAB,
  label,
  otus_tax,
  epithelium,
  today
)
}
\arguments{
\item{sgcca.centroid}{SGCCA output}

\item{STAB}{Output of bootstrap by \code{boot_sgcca} function.}

\item{label}{Name of the output files}

\item{otus_tax}{matrix as output of \code{taxonomy} function}

\item{epithelium}{Data from a file from the lab}

\item{today}{Date as in character format}
}
\description{
Looks if the score is really associated in the permutations
Does the ORA enrichment and fgsea analysis on the genes
}
