% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integration}
\alias{integration}
\title{Integrate}
\usage{
integration(A, meta, label)
}
\arguments{
\item{A}{The list with the three datsets to integrate}

\item{meta}{The metadata}

\item{label}{A label to apply to the files}
}
\value{
The sgcca output
}
\description{
integrates and does the bootstrap of the data provided. It creates also the
pdfs and the files with data
}
