% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{norm_otus}
\alias{norm_otus}
\title{Normalize OTUS}
\usage{
norm_otus(otus_tax_i, otus_table_i)
}
\arguments{
\item{otus_tax_i}{Taxonomic table}

\item{otus_table_i}{OTUs table}
}
\value{
A normalized and filtered matrix
}
\description{
Uses metagenomeSeq
}
