% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanovas.R
\name{permanova_expr}
\alias{permanova_expr}
\title{Permanova}
\usage{
permanova_expr(expr, meta)
}
\arguments{
\item{expr}{Matrix of the expression}

\item{meta}{Data frame with the variables to compare}
}
\value{
An anova.cca object
}
\description{
Permanova
}
