% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Test prevalence}
\usage{
prevalence(presence, absence)
}
\arguments{
\item{presence}{Matrix with presence of certain microorganism}

\item{absence}{Matrix with the absence of certain microorganism}
}
\description{
Iterate in the tables to reach
}
