% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_RGCCA.R
\name{wilks_rgcca}
\alias{wilks_rgcca}
\title{Check the efficacy of RGCCA}
\usage{
wilks_rgcca(a, rgcca)
}
\arguments{
\item{a}{The value}

\item{rgcca}{The output of \code{\link[RGCCA]{sgcca}} or
\code{\link[RGCCA]{rgcca}}}
}
\description{
This function test some help from
\url{https://onlinecourses.science.psu.edu/stat505/node/68}
Performs the wilks test on the model
}
